+++ server/helpers/timestamp.js
@@ -1,12 +1,13 @@
-const timestamp = (divider = "-", date) => {
+const timestamp = (date, divider = "-:") => {
     let today = date ? new Date(date) : new Date();
     let d = String(today.getDate()).padStart(2, '0');
     let m = String(today.getMonth() + 1).padStart(2, '0'); //January is 0!
     let y = today.getFullYear();
     let H = String(today.getHours()).padStart(2, '0');
     let M = String(today.getMinutes()).padStart(2, '0');
+    let [sd = '-', st = ':'] = divider.split("")
 
-    return `${d}${divider}${m}${divider}${y} ${H}:${M}`;
+    return `${d}${sd}${m}${sd}${y} ${H}${st}${M}`;
 }
 
 const secondsToTime = (s) => {